rename utils/geospatial -> utils/geometry-algorithms
